requirements.txt
fastapi
uvicorn
streamlit
scikit-learn
pandas
numpy
joblib
xgboost
pydantic
python-multipart
python-dotenv